refactor: consolidate common DTOs into singular dto/ directory
Merge src/common/dtos/ and src/common/dto/ into a single convention,
keeping the singular dto/ folder per NestJS convention.

- Move create-health-task.dto.ts, update-health-task.dto.ts,
  update-profile.dto.ts from dtos/ to dto/
- Merge pagination.dto.ts: combine PaginationDto (with @Transform
  defaults) with SortOrder, SortFieldDto, PaginationMetaDto, and
  meta-based PaginatedResponseDto from dtos/ version
- Keep dto/paginated-response.dto.ts unchanged (flat constructor
  based PaginatedResponseDto)
- Update all 17 imports across 8 files to use dto/ paths
- Delete dtos/ folder including README.md

Closes #795
