+++ src/modules/users/dto/user-filter.dto.ts
@@ -2,7 +2,7 @@ import { Type } from 'class-transformer';
 import { IsOptional, IsEnum, IsBoolean, IsDateString, IsString, IsArray } from 'class-validator';
 import { ApiPropertyOptional } from '@nestjs/swagger';
 import { Role } from '@modules/auth/enums/role.enum';
-import { PaginationDto } from '../../../common/dtos/pagination.dto';
+import { PaginationDto } from '../../../common/dto/pagination.dto';
 
 export class UserFilterDto extends PaginationDto {
   @ApiPropertyOptional({
