+++ src/modules/users/dto/user-search.dto.ts
@@ -1,7 +1,7 @@
 import { Type, Transform } from 'class-transformer';
 import { IsOptional, IsString, IsEnum, IsInt, Min, Max, IsBoolean, IsDateString } from 'class-validator';
 import { ApiPropertyOptional, ApiProperty } from '@nestjs/swagger';
-import { SortOrder } from '../../../common/dtos/pagination.dto';
+import { SortOrder } from '../../../common/dto/pagination.dto';
 import { Role } from '@modules/auth/enums/role.enum';
 import { UserStatus } from '@modules/auth/enums/user-status.enum';
