+++ src/modules/users/services/user-search.service.ts
@@ -5,7 +5,7 @@ import { User } from '../../../entities/user.entity';
 import { UserSearchDto, UserSearchResponseDto, UserSearchResultDto } from '../dto/user-search.dto';
 import { Role } from '@modules/auth/enums/role.enum';
 import { UserStatus } from '@modules/auth/enums/user-status.enum';
-import { SortOrder } from '../../../common/dtos/pagination.dto';
+import { SortOrder } from '../../../common/dto/pagination.dto';
 
 @Injectable()
 export class UserSearchService {
