+++ src/modules/users/users-profile.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { UpdateProfileDto, ProfileResponseDto } from '../../common/dtos/update-profile.dto';
+import { UpdateProfileDto, ProfileResponseDto } from '../../common/dto/update-profile.dto';
 import { User } from '../../entities/user.entity';
 import { Role } from '@modules/auth/enums/role.enum';
 import { UserStatus } from '@modules/auth/enums/user-status.enum';
